+++ jax/_src/lax/convolution.py
@@ -818,10 +818,10 @@ def _conv_general_dilated_lower(
         dimension_numbers=dnums,
         feature_group_count=mlir.i64_attr(feature_group_count),
         batch_group_count=mlir.i64_attr(batch_group_count),
-        window_strides=mlir.dense_int_array(window_strides),
-        padding=mlir.dense_int_elements(padding),
-        lhs_dilation=mlir.dense_int_array(lhs_dilation),
-        rhs_dilation=mlir.dense_int_array(rhs_dilation),
+        window_strides=mlir.i64_array_attr(window_strides),
+        padding=mlir.i64_array_attr(padding),
+        lhs_dilation=mlir.i64_array_attr(lhs_dilation),
+        rhs_dilation=mlir.i64_array_attr(rhs_dilation),
         window_reversal=window_reversal,
         precision_config=lax.precision_attr(precision))
     return [mlir.lower_with_sharding_in_types(ctx, out, aval_out)]
@@ -843,9 +843,9 @@ def _conv_general_dilated_lower(
           dimension_numbers=dnums,
           feature_group_count=mlir.i64_attr(feature_group_count),
           batch_group_count=mlir.i64_attr(batch_group_count),
-          window_strides=mlir.dense_int_array(window_strides),
-          lhs_dilation=mlir.dense_int_array(lhs_dilation),
-          rhs_dilation=mlir.dense_int_array(rhs_dilation),
+          window_strides=mlir.i64_array_attr(window_strides),
+          lhs_dilation=mlir.i64_array_attr(lhs_dilation),
+          rhs_dilation=mlir.i64_array_attr(rhs_dilation),
           window_reversal=window_reversal,
           precision_config=lax.precision_attr(precision))
     ]