+++ jax/_src/pallas/triton/lowering.py
@@ -1318,7 +1318,7 @@ def debug_print_lowering_rule(
       f" {fmt} ",
       hex=False,
       args=args,
-      is_signed=ir.DenseI32ArrayAttr.get([
+      is_signed=mlir.i32_array_attr([
           jnp.issubdtype(aval.dtype, jnp.signedinteger) for aval in ctx.avals_in
       ]),
   )
@@ -1344,11 +1344,7 @@ def _set_attr(v: ir.Value, name: str, attr: ir.Attribute) -> None:
 def _multiple_of_rule(ctx: LoweringRuleContext, x, values: Sequence[int]):
   [x_aval] = ctx.avals_in
   assert max(1, len(x_aval.shape)) == len(values)
-  _set_attr(
-      x,
-      "tt.divisibility",
-      ir.DenseIntElementsAttr.get(np.asarray(values, dtype=np.int32)),
-  )
+  _set_attr(x, "tt.divisibility", mlir.i32_array_attr(values))
   return x